+++ src/django_q_registry/models.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 import json
 import logging
+from collections.abc import Callable
 from typing import Any
-from typing import Callable
 
 from django.core.serializers.json import DjangoJSONEncoder
 from django.db import models